refactor(navbar): extract glow effect logic into custom hook
