+++ app/components/navbar.tsx
@@ -1,8 +1,9 @@
 'use client';
 
-import { CSSProperties, useEffect, useRef, useState } from 'react';
+import { useEffect, useState } from 'react';
 import Link from 'next/link';
 import { Menu, X, Activity } from 'lucide-react';
+import { useGlowEffect } from '@/hooks/useGlowEffect';
 
 function GithubMark() {
   return (
@@ -21,92 +22,8 @@ const NAV_LINKS = [
 
 export default function Navbar() {
   const [open, setOpen] = useState(false);
-  const shellRef = useRef<HTMLDivElement>(null);
-  const rectRef = useRef<DOMRect | null>(null);
-  const animationRef = useRef<number | null>(null);
-  const targetRef = useRef({ x: 50, y: 50 });
-  const currentRef = useRef({ x: 50, y: 50 });
-  const activeRef = useRef(false);
-
-  const shellVars = {
-    // Defaults keep the glow centered before pointer interaction.
-    ['--mx' as string]: '50%',
-    ['--my' as string]: '50%',
-    ['--glow-opacity' as string]: '0',
-    ['--border-opacity' as string]: '0',
-  } as CSSProperties & Record<string, string>;
-
-  const updateRect = () => {
-    const shell = shellRef.current;
-    rectRef.current = shell ? shell.getBoundingClientRect() : null;
-  };
-
-  const animateGlow = () => {
-    const shell = shellRef.current;
-
-    if (!shell) {
-      animationRef.current = null;
-      return;
-    }
-
-    const smoothing = 0.16;
-    currentRef.current.x += (targetRef.current.x - currentRef.current.x) * smoothing;
-    currentRef.current.y += (targetRef.current.y - currentRef.current.y) * smoothing;
-
-    shell.style.setProperty('--mx', `${currentRef.current.x}%`);
-    shell.style.setProperty('--my', `${currentRef.current.y}%`);
-    shell.style.setProperty('--glow-opacity', activeRef.current ? '1' : '0');
-    shell.style.setProperty('--border-opacity', activeRef.current ? '1' : '0');
-
-    const settled =
-      Math.abs(targetRef.current.x - currentRef.current.x) < 0.08 &&
-      Math.abs(targetRef.current.y - currentRef.current.y) < 0.08;
-
-    if (!activeRef.current && settled) {
-      animationRef.current = null;
-      return;
-    }
-
-    animationRef.current = requestAnimationFrame(animateGlow);
-  };
-
-  const startAnimation = () => {
-    if (animationRef.current !== null) return;
-    animationRef.current = requestAnimationFrame(animateGlow);
-  };
-
-  useEffect(() => {
-    updateRect();
-
-    const handleViewportChange = () => {
-      updateRect();
-    };
-
-    window.addEventListener('resize', handleViewportChange);
-    window.addEventListener('scroll', handleViewportChange, true);
-
-    let resizeObserver: ResizeObserver | null = null;
-
-    if (shellRef.current && 'ResizeObserver' in window) {
-      resizeObserver = new ResizeObserver(handleViewportChange);
-      resizeObserver.observe(shellRef.current);
-    }
-
-    return () => {
-      if (animationRef.current !== null) {
-        cancelAnimationFrame(animationRef.current);
-      }
-
-      window.removeEventListener('resize', handleViewportChange);
-      window.removeEventListener('scroll', handleViewportChange, true);
-      resizeObserver?.disconnect();
-    };
-  }, []);
-
-  const handleLogoClick = () => {
-    setOpen(false);
-    window.scrollTo({ top: 0, behavior: 'smooth' });
-  };
+  const { shellRef, shellVars, handleMouseEnter, handleMouseMove, handleMouseLeave } =
+    useGlowEffect();
 
   useEffect(() => {
     const mediaQuery = window.matchMedia('(min-width: 768px)');
@@ -133,6 +50,11 @@ export default function Navbar() {
     };
   }, []);
 
+  const handleLogoClick = () => {
+    setOpen(false);
+    window.scrollTo({ top: 0, behavior: 'smooth' });
+  };
+
   return (
     <header className="fixed inset-x-0 top-0 z-50 px-4 pt-4 sm:px-6">
       <div className="mx-auto max-w-6xl">
@@ -140,27 +62,9 @@ export default function Navbar() {
           ref={shellRef}
           className="relative overflow-hidden rounded-2xl border border-white/25 bg-black/45 backdrop-blur-xl shadow-[0_14px_40px_rgba(0,0,0,0.45)]"
           style={shellVars}
-          onMouseEnter={updateRect}
-          onMouseMove={(event) => {
-            if (!rectRef.current) {
-              rectRef.current = event.currentTarget.getBoundingClientRect();
-            }
-
-            const rect = rectRef.current;
-
-            if (!rect) return;
-            const x = ((event.clientX - rect.left) / rect.width) * 100;
-            const y = ((event.clientY - rect.top) / rect.height) * 100;
-
-            targetRef.current = { x, y };
-            activeRef.current = true;
-            startAnimation();
-          }}
-          onMouseLeave={() => {
-            activeRef.current = false;
-            rectRef.current = null;
-            startAnimation();
-          }}
+          onMouseEnter={handleMouseEnter}
+          onMouseMove={handleMouseMove}
+          onMouseLeave={handleMouseLeave}
         >
           <div
             className="pointer-events-none absolute inset-0 transition-opacity duration-300 ease-out"
